fix: change datatype of price product in product table

+++ database/product.sql
@@ -1,7 +1,7 @@
 CREATE TABLE products(
     id SERIAL,
     name VARCHAR(50) NOT NULL,
-    price DOUBLE NOT NULL,
+    price FLOAT NOT NULL,
     stock INT NOT NULL,
     PRIMARY KEY(id)
 );
